% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{value_iteration}
\alias{value_iteration}
\title{Do value iteration for GridWorld}
\usage{
value_iteration(reward, obstacles, end_cell, wind, beta)
}
\description{
Do value iteration for GridWorld
}
