% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{value_iteration_soft}
\alias{value_iteration_soft}
\title{Do value iteration for GridWorld using softmax}
\usage{
value_iteration_soft(reward, obstacles, end_cell)
}
\description{
Do value iteration for GridWorld using softmax
}
